rbd: consolidate error types & values in an errors.go file
